Add files via upload
The "Shark Tank Analysis" repository on GitHub contains a project that analyzes the success rate of entrepreneurs who pitch their business ideas on the popular TV show "Shark Tank". The project uses data analysis techniques by using Mysql to explore patterns and trends in the show's history. A PowerPoint presentation is also included to summarize the findings in a clear and concise manner.