Add files via upload
The "Swiggy Analysis" repository on GitHub contains a project that analyzes the business model, growth, and revenue of the Indian food delivery company Swiggy. The project uses data analysis by using Mysql to explore Swiggy's performance in the market, its customer base, and its competition. A PowerPoint presentation is included to present the findings in a clear and concise manner